Add files via upload
The data used in this analysis was sourced from the 1994 Census Bureau database, and it was originally collected by Ronny Kohavi and Barry Becker as part of their work in Data Mining and Visualization at Silicon Graphics. The primary objective of this analysis is to ascertain whether an individual's annual income exceeds $50,000.